🛠️(refactor) DIManager를 통해 객체들 생성하여 register만 관리
